# speed up `ensure_parsnip_format()`