A helper that gets a good bit of traffic at `predict()`.

With `main` dev:

``` r
library(parsnip)

lm_fit <- fit(linear_reg(), mpg ~ ., mtcars)

bench::mark(old = predict(lm_fit, mtcars))
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old           619µs    654µs     1476.    1.08MB     35.9
```

With this PR:

``` r
bench::mark(new = predict(lm_fit, mtcars))
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new           389µs    404µs     2419.     599KB     40.3
```

<sup>Created on 2023-04-02 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>